ref: Move Signed extractor to extractors mod
